Delete Pessoa rodando

+++ back/src/main/java/br/com/cesrc/seusuas/service/PessoaService.java
@@ -9,19 +9,25 @@ import java.util.List;
 @Service
 public class PessoaService {
     private final PessoaRepository pessoaRepository;
+
     public PessoaService(PessoaRepository pessoaRepository){
         this.pessoaRepository = pessoaRepository;
     }
+
     public PessoaModel salvarPessoa (PessoaModel pessoa){
         return pessoaRepository.save(pessoa);
     }
-    public void deletarAcolhido (Long id){
+    public void deletarPessoa(Long id) {
+        if (!pessoaRepository.existsById(id)) {
+            throw new RuntimeException("Pessoa com ID " + id + " não encontrada");
+        }
         pessoaRepository.deleteById(id);
     }
-    public List<PessoaModel> listarAcolhido(){
+
+    public List<PessoaModel> listarPessoa(){
         return pessoaRepository.findAll();
     }
-    public void editarAcolhido(Long id, PessoaRequest pessoaRequest){
+    public void editarPessoa(Long id, PessoaRequest pessoaRequest){
         PessoaModel PessoaExistente = pessoaRepository.findById(id).orElseThrow(() -> new RuntimeException(id +" não existe"));
         PessoaExistente.setNome(pessoaRequest.getNome());
         PessoaExistente.setDataNascimento(pessoaRequest.getDataNascimento());
